added self-test fro DotExporter
