+++ pepper-framework/src/main/java/org/corpus_tools/pepper/modules/coreModules/DOTExporter.java
@@ -21,6 +21,7 @@ import java.util.List;
 
 import org.corpus_tools.pepper.common.DOCUMENT_STATUS;
 import org.corpus_tools.pepper.common.PepperConfiguration;
+import org.corpus_tools.pepper.core.SelfTestDesc;
 import org.corpus_tools.pepper.impl.PepperExporterImpl;
 import org.corpus_tools.pepper.impl.PepperMapperImpl;
 import org.corpus_tools.pepper.modules.PepperMapper;
@@ -33,15 +34,27 @@ import org.osgi.service.component.annotations.Component;
 
 @Component(name = "DOTExporterComponent", factory = "PepperExporterComponentFactory", enabled = true)
 public class DOTExporter extends PepperExporterImpl {
+	public static final String FORMAT_NAME = "dot";
+	public static final String FORMAT_VERSION= "1.0";
+
 	public DOTExporter() {
 		super("DOTExporter");
 		setSupplierContact(URI.createURI(PepperConfiguration.EMAIL));
 		setSupplierHomepage(URI.createURI(PepperConfiguration.HOMEPAGE));
 		setDesc("This exporter exports a Salt model to the dot syntax. This can be used to create a graphical representation of the Salt model. ");
 		// set list of formats supported by this module
-		this.addSupportedFormat("dot", "1.0", null);
+		this.addSupportedFormat(FORMAT_NAME, FORMAT_VERSION, null);
 	}
 
+	@Override
+	public SelfTestDesc getSelfTestDesc() {
+		return new SelfTestDesc(
+				getResources().appendSegment("modules").appendSegment("selfTests").appendSegment("dotExporter")
+						.appendSegment("in"),
+				getResources().appendSegment("modules").appendSegment("selfTests").appendSegment("dotExporter")
+						.appendSegment("expected"));
+	}
+	
 	@Override
 	public PepperMapper createPepperMapper(Identifier sElementId) {
 		PepperMapper mapper = new PepperMapperImpl() {
@@ -83,7 +96,8 @@ public class DOTExporter extends PepperExporterImpl {
 			segments.append("/");
 			segments.append(segment);
 		}
-		outputURI = URI.createFileURI(this.getCorpusDesc().getCorpusPath().toFileString() + segments.toString() + "." + SaltUtil.FILE_ENDING_SALT_XML);
+		outputURI = URI.createFileURI(this.getCorpusDesc().getCorpusPath().toFileString() + segments.toString() + "."
+				+ SaltUtil.FILE_ENDING_SALT_XML);
 
 		mapper.setResourceURI(outputURI);
 		return (mapper);
