+++ pepper-framework/src/main/resources/modules/selfTests/dotExporter/expected/rootCorpus.dot
@@ -1,0 +1,16 @@
+digraph G {
+ordering=out;
+<salt:/rootCorpus>[shape=Mrecord,color=gray28,style=filled,label="{{salt:/rootCorpus}|{name= rootCorpus}}"];
+<salt:/rootCorpus/subCorpus1>[shape=Mrecord,color=gray28,style=filled,label="{{salt:/rootCorpus/subCorpus1}|{name= subCorpus1}}"];
+<salt:/rootCorpus> -> <salt:/rootCorpus/subCorpus1>[color=gray28,style=filled,label="name= corpRel1"];
+<salt:/rootCorpus/subCorpus1/doc1>[shape=Mrecord,color=gray,style=filled,label="{{salt:/rootCorpus/subCorpus1/doc1}|{name= doc1}}"];
+<salt:/rootCorpus/subCorpus1> -> <salt:/rootCorpus/subCorpus1/doc1>[color=gray,style=filled,label="name= corpDocRel1"];
+<salt:/rootCorpus/subCorpus1/doc2>[shape=Mrecord,color=gray,style=filled,label="{{salt:/rootCorpus/subCorpus1/doc2}|{name= doc2}}"];
+<salt:/rootCorpus/subCorpus1> -> <salt:/rootCorpus/subCorpus1/doc2>[color=gray,style=filled,label="name= corpDocRel2"];
+<salt:/rootCorpus/subCorpus2>[shape=Mrecord,color=gray28,style=filled,label="{{salt:/rootCorpus/subCorpus2}|{name= subCorpus2}}"];
+<salt:/rootCorpus> -> <salt:/rootCorpus/subCorpus2>[color=gray28,style=filled,label="name= corpRel2"];
+<salt:/rootCorpus/subCorpus2/doc3>[shape=Mrecord,color=gray,style=filled,label="{{salt:/rootCorpus/subCorpus2/doc3}|{name= doc3}}"];
+<salt:/rootCorpus/subCorpus2> -> <salt:/rootCorpus/subCorpus2/doc3>[color=gray,style=filled,label="name= corpDocRel3"];
+<salt:/rootCorpus/subCorpus2/doc4>[shape=Mrecord,color=gray,style=filled,label="{{salt:/rootCorpus/subCorpus2/doc4}|{name= doc4}}"];
+<salt:/rootCorpus/subCorpus2> -> <salt:/rootCorpus/subCorpus2/doc4>[color=gray,style=filled,label="name= corpDocRel4"];
+}
