+++ pepper-framework/src/main/resources/modules/selfTests/saltValidator/in/saltProject.salt
@@ -1,0 +1,62 @@
+<?xml version='1.0' encoding='UTF-8'?>
+<saltCommon:SaltProject xmlns:sCorpusStructure="sCorpusStructure" xmlns:xmi="http://www.omg.org/XMI" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:saltCore="saltCore" xmlns:saltCommon="saltCommon" xmi:version="2.0">
+	<sCorpusGraphs>
+	<labels xsi:type="saltCore:SElementId" namespace="salt" name="id" value="T::corpusGraph1"/>
+	<nodes xsi:type="sCorpusStructure:SCorpus">
+		<labels xsi:type="saltCore:SElementId" namespace="salt" name="id" value="T::salt:/rootCorpus"/>
+		<labels xsi:type="saltCore:SFeature" namespace="salt" name="SNAME" value="T::rootCorpus"/>
+	</nodes>
+	<nodes xsi:type="sCorpusStructure:SCorpus">
+		<labels xsi:type="saltCore:SElementId" namespace="salt" name="id" value="T::salt:/rootCorpus/subCorpus1"/>
+		<labels xsi:type="saltCore:SFeature" namespace="salt" name="SNAME" value="T::subCorpus1"/>
+	</nodes>
+	<nodes xsi:type="sCorpusStructure:SCorpus">
+		<labels xsi:type="saltCore:SElementId" namespace="salt" name="id" value="T::salt:/rootCorpus/subCorpus2"/>
+		<labels xsi:type="saltCore:SFeature" namespace="salt" name="SNAME" value="T::subCorpus2"/>
+	</nodes>
+	<nodes xsi:type="sCorpusStructure:SDocument">
+		<labels xsi:type="saltCore:SElementId" namespace="salt" name="id" value="T::salt:/rootCorpus/subCorpus1/doc1"/>
+		
+		<labels xsi:type="saltCore:SFeature" namespace="salt" name="SNAME" value="T::doc1"/>
+	</nodes>
+	<nodes xsi:type="sCorpusStructure:SDocument">
+		<labels xsi:type="saltCore:SElementId" namespace="salt" name="id" value="T::salt:/rootCorpus/subCorpus1/doc2"/>
+		
+		<labels xsi:type="saltCore:SFeature" namespace="salt" name="SNAME" value="T::doc2"/>
+	</nodes>
+	<nodes xsi:type="sCorpusStructure:SDocument">
+		<labels xsi:type="saltCore:SElementId" namespace="salt" name="id" value="T::salt:/rootCorpus/subCorpus2/doc3"/>
+		
+		<labels xsi:type="saltCore:SFeature" namespace="salt" name="SNAME" value="T::doc3"/>
+	</nodes>
+	<nodes xsi:type="sCorpusStructure:SDocument">
+		<labels xsi:type="saltCore:SElementId" namespace="salt" name="id" value="T::salt:/rootCorpus/subCorpus2/doc4"/>
+		
+		<labels xsi:type="saltCore:SFeature" namespace="salt" name="SNAME" value="T::doc4"/>
+	</nodes>
+	<edges xsi:type="sCorpusStructure:SCorpusRelation" source="//@nodes.0" target="//@nodes.1">
+		<labels xsi:type="saltCore:SElementId" namespace="salt" name="id" value="T::salt:/corpRel1"/>
+		<labels xsi:type="saltCore:SFeature" namespace="salt" name="SNAME" value="T::corpRel1"/>
+	</edges>
+	<edges xsi:type="sCorpusStructure:SCorpusRelation" source="//@nodes.0" target="//@nodes.2">
+		<labels xsi:type="saltCore:SElementId" namespace="salt" name="id" value="T::salt:/corpRel2"/>
+		<labels xsi:type="saltCore:SFeature" namespace="salt" name="SNAME" value="T::corpRel2"/>
+	</edges>
+	<edges xsi:type="sCorpusStructure:SCorpusDocumentRelation" source="//@nodes.1" target="//@nodes.3">
+		<labels xsi:type="saltCore:SElementId" namespace="salt" name="id" value="T::salt:/corpDocRel1"/>
+		<labels xsi:type="saltCore:SFeature" namespace="salt" name="SNAME" value="T::corpDocRel1"/>
+	</edges>
+	<edges xsi:type="sCorpusStructure:SCorpusDocumentRelation" source="//@nodes.1" target="//@nodes.4">
+		<labels xsi:type="saltCore:SElementId" namespace="salt" name="id" value="T::salt:/corpDocRel2"/>
+		<labels xsi:type="saltCore:SFeature" namespace="salt" name="SNAME" value="T::corpDocRel2"/>
+	</edges>
+	<edges xsi:type="sCorpusStructure:SCorpusDocumentRelation" source="//@nodes.2" target="//@nodes.5">
+		<labels xsi:type="saltCore:SElementId" namespace="salt" name="id" value="T::salt:/corpDocRel3"/>
+		<labels xsi:type="saltCore:SFeature" namespace="salt" name="SNAME" value="T::corpDocRel3"/>
+	</edges>
+	<edges xsi:type="sCorpusStructure:SCorpusDocumentRelation" source="//@nodes.2" target="//@nodes.6">
+		<labels xsi:type="saltCore:SElementId" namespace="salt" name="id" value="T::salt:/corpDocRel4"/>
+		<labels xsi:type="saltCore:SFeature" namespace="salt" name="SNAME" value="T::corpDocRel4"/>
+	</edges>
+	</sCorpusGraphs>
+</saltCommon:SaltProject>
